extract to files - precommit hook now passing
