+++ tests/unit/monitor/alerts/group_alerts_by_table/test_slack_alerts_group_by.py
@@ -1,14 +1,8 @@
-import json
-from dataclasses import dataclass
-from datetime import datetime
-from typing import Dict, List, Optional, Union
-
 from parametrization import Parametrization
 
 from elementary.monitor.alerts.group_of_alerts import (
     ErrorComponent,
     FailureComponent,
-    GroupOfAlerts,
     GroupOfAlertsByAll,
     GroupOfAlertsBySingleAlert,
     GroupOfAlertsByTable,
@@ -20,167 +14,26 @@ from elementary.monitor.alerts.group_of_alerts import (
 from elementary.monitor.data_monitoring.data_monitoring_alerts import (
     DataMonitoringAlerts,
 )
-from elementary.monitor.fetchers.alerts.normalized_alert import CHANNEL_KEY
-
-# what I use
-# self.config.slack_group_alerts_by
-# self.config.slack_channel_name
-# self.execution_properties["had_group_by_all"] = (len(alls_group) > 0)
-# self.execution_properties["had_group_by_table"] = (len(by_table_group) > 0)
-# self.execution_properties["had_group_by_alert"] = (len(by_alert_group) > 0)
-# al.slack_group_alerts_by
-# al.model_unique_id
-
-
-#####################
-### Mock Classes ####
-#####################
-@dataclass
-class MockAlert:
-    status: Optional[str]  # should be in "warn", "error" , "fail[ure]"
-    slack_group_alerts_by: Optional[str]
-    model_unique_id: Optional[str]
-    slack_channel: Optional[str]
-    detected_at: Optional[Union[str, datetime]]
-    model_meta: Optional[
-        str
-    ]  # this string should be a json of a dict that has or has not the key "channel"
-    owners: Optional[List[str]]
-    subscribers: Optional[List[str]]
-    tags: Optional[str]
-
-
-@dataclass
-class MockConfig:
-    slack_group_alerts_by: Optional[str]
-    slack_channel_name: Optional[str]
-
-
-@dataclass
-class MockDataMonitoringAlerts:
-    config: MockConfig
-    execution_properties: Dict
-
-
-def mock_data_monitoring_alerts(mock_config):
-    return MockDataMonitoringAlerts(config=mock_config, execution_properties=dict())
-
-
-###################
-### Mock Data  ####
-###################
-DEFAULT_CHANNEL = "roi-playground"
-OTHER_CHANNEL = "roi-playground-2"
-MODEL_1 = "models.bla.model1"
-MODEL_2 = "models.bla.model2"
-MODEL_3 = "models.bla.model3"
-OWNER_1 = "owner1@elementary-data.com"
-OWNER_2 = "owner2@elementary-data.com"
-OWNER_3 = "owner3@elementary-data.com"
-TAGS_1 = "marketing, finance"
-TAGS_2 = "finance, data-ops"
-TAGS_3 = "marketing, finance, data-ops"
-DETECTED_AT = "1992-11-11 20:00:03+0200"
-EMPTY_DICT = json.dumps(dict())
-OTHER_CHANNEL_IN_DICT = json.dumps({CHANNEL_KEY: OTHER_CHANNEL})
-
-AL_WARN_MODEL1_NO_CHANNEL_NO_GROUPING = MockAlert(
-    status="warn",
-    slack_group_alerts_by=None,
-    model_unique_id=MODEL_1,
-    slack_channel=None,
-    detected_at=DETECTED_AT,
-    model_meta=EMPTY_DICT,
-    owners=[OWNER_1, OWNER_2],
-    subscribers=[],
-    tags=TAGS_1,
+from tests.unit.monitor.alerts.group_alerts_by_table.mock_classes import MockConfig
+from tests.unit.monitor.alerts.group_alerts_by_table.mock_data import (
+    AL_ERROR_MODEL2_NO_CHANNEL_WITH_GROUPING_BY_TABLE,
+    AL_ERROR_MODEL3_NO_CHANNEL_WITH_GROUPING_BY_ALERT,
+    AL_FAIL_MODEL1_WITH_CHANNEL_NO_GROUPING,
+    AL_FAIL_MODEL2_NO_CHANNEL_NO_GROUPING,
+    AL_FAIL_MODEL2_NO_CHANNEL_WITH_GROUPING_BY_ALERT,
+    AL_FAIL_MODEL2_NO_CHANNEL_WITH_GROUPING_BY_TABLE,
+    AL_WARN_MODEL1_NO_CHANNEL_NO_GROUPING,
+    DEFAULT_CHANNEL,
+    OTHER_CHANNEL,
+    OWNER_1,
+    OWNER_2,
+    OWNER_3,
+    TAGS_3,
 )
-
-AL_FAIL_MODEL1_WITH_CHANNEL_NO_GROUPING = MockAlert(
-    status="fail",
-    slack_group_alerts_by=None,
-    model_unique_id=MODEL_1,
-    slack_channel=OTHER_CHANNEL,
-    detected_at=DETECTED_AT,
-    model_meta="{}",
-    owners=[OWNER_1, OWNER_3],
-    subscribers=[],
-    tags=TAGS_2,
+from tests.unit.monitor.alerts.group_alerts_by_table.utils import (
+    check_eq_group_alerts,
+    mock_data_monitoring_alerts,
 )
-
-AL_FAIL_MODEL2_NO_CHANNEL_NO_GROUPING = MockAlert(
-    status="fail",
-    slack_group_alerts_by=None,
-    model_unique_id=MODEL_2,
-    slack_channel=None,
-    detected_at=DETECTED_AT,
-    model_meta=EMPTY_DICT,
-    owners=[OWNER_1],
-    subscribers=[],
-    tags=TAGS_3,
-)
-AL_FAIL_MODEL2_NO_CHANNEL_WITH_GROUPING_BY_ALERT = MockAlert(
-    status="fail",
-    slack_group_alerts_by="by_alert",
-    model_unique_id=MODEL_2,
-    slack_channel=None,
-    detected_at=DETECTED_AT,
-    model_meta=EMPTY_DICT,
-    owners=[OWNER_1],
-    subscribers=[],
-    tags=TAGS_1,
-)
-AL_FAIL_MODEL2_NO_CHANNEL_WITH_GROUPING_BY_TABLE = MockAlert(
-    status="fail",
-    slack_group_alerts_by="by_table",
-    model_unique_id=MODEL_2,
-    slack_channel=None,
-    detected_at=DETECTED_AT,
-    model_meta=OTHER_CHANNEL_IN_DICT,
-    owners=[OWNER_1],
-    subscribers=[],
-    tags=TAGS_1,
-)
-AL_ERROR_MODEL2_NO_CHANNEL_WITH_GROUPING_BY_TABLE = MockAlert(
-    status="error",
-    slack_group_alerts_by="by_table",
-    model_unique_id=MODEL_2,
-    slack_channel=None,
-    detected_at=DETECTED_AT,
-    model_meta=EMPTY_DICT,
-    owners=[OWNER_1],
-    subscribers=[],
-    tags=TAGS_1,
-)
-AL_ERROR_MODEL3_NO_CHANNEL_WITH_GROUPING_BY_ALERT = MockAlert(
-    status="error",
-    slack_group_alerts_by="by_alert",
-    model_unique_id=MODEL_3,
-    slack_channel=None,
-    detected_at=DETECTED_AT,
-    model_meta=EMPTY_DICT,
-    owners=[OWNER_1],
-    subscribers=[],
-    tags=TAGS_1,
-)
-
-
-#############
-### Utils ###
-#############
-def check_eq_group_alerts(grp1: GroupOfAlerts, grp2: GroupOfAlerts):
-    ret = grp1.alerts == grp2.alerts  # same alerts
-    ret &= grp1._components_to_alerts == grp2._components_to_alerts  # same mappings
-    ret &= (
-        grp1._components_to_attn_required == grp2._components_to_attn_required
-    )  # same owners, tags, subscribers
-
-    return ret
-
-
-#################
-## Test Cases ###
-#################
 
 
 @Parametrization.autodetect_parameters()
